refactor types into their own file
