+++ src/components/Canvas/Road.tsx
@@ -3,7 +3,8 @@ import { Arrow, Group } from 'react-konva';
 import { KonvaEventObject } from 'konva/lib/Node';
 
 import { centerlineColor, highlightColor, roadColor } from '~/colors';
-import { Edge, useNetworkStore } from '~/zustand/useNetworkStore';
+import { Edge } from '~/types/Network';
+import { useNetworkStore } from '~/zustand/useNetworkStore';
 import { useSelector } from '~/zustand/useSelected';
 
 interface RoadProps {
